Move context macros to a separate modules
